Refactor codebase into modules
