+++ .eslintrc
@@ -25,7 +25,8 @@
     "no-continue": 0,
     "no-cond-assign": 0,
     "no-plusplus": 0,
-    "no-use-before-define": 0
+    "no-use-before-define": 0,
+    "lines-between-class-members": 0
   },
   "settings": {
     "flowtype": {
