+++ src/index.js
@@ -1,0 +1,19 @@
+import Tez from './tez';
+import crypto from './crypto';
+import forge from './forge';
+import utility from './utility';
+import ledger from './ledger';
+
+// tez.js Base Modules
+export {
+  crypto,
+  forge,
+  utility,
+  ledger,
+};
+
+// Miscellaneous Modules
+export { Contract } from './modules';
+
+// tez.js Library
+export default Tez;
