+++ src/modules/index.js
@@ -1,0 +1,5 @@
+import Contract from './Contract';
+
+export default {
+  Contract,
+};
