+++ src/utility.js
@@ -1,0 +1,287 @@
+// @flow
+import bs58check from 'bs58check';
+import { BigNumber } from 'bignumber.js';
+
+import type { Utility } from './types';
+
+if (typeof Buffer === 'undefined') {
+  Buffer = require('buffer'); // eslint-disable-line
+}
+
+const utility: Utility = {};
+
+utility.textEncode = (value: string): Uint8Array => new Uint8Array(Buffer.from(value, 'utf8'));
+utility.textDecode = (buffer: Uint8Array) => Buffer.from(buffer).toString('utf8');
+
+/**
+ * @description Convert from base58 to integer
+ * @param {String} v The b58 value
+ * @returns {String} The converted b58 value
+ */
+utility.b582int = (v: string): string => {
+  let rv = new BigNumber(0);
+  const alpha = '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz';
+  for (let i = 0; i < v.length; i++) {
+    rv = rv.plus(new BigNumber(alpha.indexOf(v[v.length - 1 - i])).multipliedBy(new BigNumber(alpha.length).exponentiatedBy(i)));
+  }
+  return rv.toString(16);
+};
+
+/**
+ * @description Convert from mutez to tez
+ * @param {Number} mutez The amount in mutez to convert to tez
+ * @returns {Number} The mutez amount converted to tez
+ */
+utility.totez = (mutez: number): number => parseInt(mutez, 10) / 1000000;
+
+/**
+ * @description Convert from tez to mutez
+ * @param {Number} tez The amount in tez to convert to mutez
+ * @returns {String} The tez amount converted to mutez
+ */
+utility.mutez = (tez: number): string => new BigNumber(new BigNumber(tez).toFixed(6)).multipliedBy(1000000).toString();
+
+/**
+ * @description Base58 encode
+ * @param {String} payload The value to encode
+ * @param {Object} prefixArg The Uint8Array prefix values
+ * @returns {String} The base58 encoded value
+ */
+utility.b58cencode = (payload: string, prefixArg: Uint8Array): string => {
+  const n = new Uint8Array(prefixArg.length + payload.length);
+  n.set(prefixArg);
+  // $FlowFixMe
+  n.set(payload, prefixArg.length);
+  // $FlowFixMe
+  return bs58check.encode(Buffer.from(n, 'hex'));
+};
+
+/**
+ * @description Base58 decode
+ * @param {String} payload The value to decode
+ * @param {Object} prefixArg The Uint8Array prefix values
+ * @returns {String} The decoded base58 value
+ */
+utility.b58cdecode = (enc: string, prefixArg: Uint8Array): string => bs58check.decode(enc).slice(prefixArg.length);
+
+/**
+ * @description Buffer to hex
+ * @param {Object} buffer The buffer to convert to hex
+ * @returns {String} Converted hex value
+ */
+utility.buf2hex = (buffer: Uint8Array | string): string => {
+  // $FlowFixMe
+  const byteArray = new Uint8Array(buffer);
+  const hexParts = [];
+  byteArray.forEach((byte) => {
+    const hex = byte.toString(16);
+    const paddedHex = (`00${hex}`).slice(-2);
+    hexParts.push(paddedHex);
+  });
+  return hexParts.join('');
+};
+
+/**
+ * @description Hex to Buffer
+ * @param {String} hex The hex to convert to buffer
+ * @returns {Object} Converted buffer value
+ */
+utility.hex2buf = (hex: string): Uint8Array => (
+  // $FlowFixMe
+  new Uint8Array(hex.match(/[\da-f]{2}/gi).map(h => parseInt(h, 16)))
+);
+
+/**
+ * @description Generate a hex nonce
+ * @param {Number} length The length of the nonce
+ * @returns {String} The nonce of the given length
+ */
+utility.hexNonce = (length: number): string => {
+  const chars = '0123456789abcedf';
+  let hex = '';
+  while (length--) {
+    hex += chars[(Math.random() * 16) | 0];
+  }
+  return hex;
+};
+
+/**
+ * @description Merge two buffers together
+ * @param {Object} b1 The first buffer
+ * @param {Object} b2 The second buffer
+ * @returns {Object} The merged buffer
+ */
+utility.mergebuf = (b1: Uint8Array, b2: Uint8Array): Uint8Array => {
+  const r = new Uint8Array(b1.length + b2.length);
+  r.set(b1);
+  r.set(b2, b1.length);
+  return r;
+};
+
+utility.sexp2mic = function me(mi: string): * {
+  mi = mi.replace(/(?:@[a-z_]+)|(?:#.*$)/mg, '')
+    .replace(/\s+/g, ' ')
+    .trim();
+  if (mi.charAt(0) === '(') mi = mi.slice(1, -1);
+  let pl = 0;
+  let sopen = false;
+  let escaped = false;
+  const ret = {
+    prim: '',
+    args: [],
+  };
+  let val = '';
+  for (let i = 0; i < mi.length; i++) {
+    if (escaped) {
+      val += mi[i];
+      escaped = false;
+      continue;
+    } else if ((i === (mi.length - 1) && sopen === false) || (mi[i] === ' ' && pl === 0 && sopen === false)) {
+      if (i === (mi.length - 1)) val += mi[i];
+      if (val) {
+        if (val === parseInt(val, 10).toString()) {
+          if (!ret.prim) return { int: val };
+          ret.args.push({ int: val });
+        } else if (val[0] === '0') {
+          if (!ret.prim) return { bytes: val };
+          ret.args.push({ bytes: val });
+        } else if (ret.prim) {
+          ret.args.push(me(val));
+        } else {
+          ret.prim = val;
+        }
+        val = '';
+      }
+      continue;
+    } else if (mi[i] === '"' && sopen) {
+      sopen = false;
+      if (!ret.prim) return { string: val };
+      ret.args.push({ string: val });
+      val = '';
+      continue;
+    } else if (mi[i] === '"' && !sopen && pl === 0) {
+      sopen = true;
+      continue;
+    } else if (mi[i] === '\\') escaped = true;
+    else if (mi[i] === '(') pl++;
+    else if (mi[i] === ')') pl--;
+    val += mi[i];
+  }
+  return ret;
+};
+
+utility.mic2arr = function me2(s: any): * {
+  let ret: any = [];
+  if (Object.prototype.hasOwnProperty.call(s, 'prim')) {
+    if (s.prim === 'Pair') {
+      ret.push(me2(s.args[0]));
+      ret = ret.concat(me2(s.args[1]));
+    } else if (s.prim === 'Elt') {
+      ret = {
+        key: me2(s.args[0]),
+        val: me2(s.args[1]),
+      };
+    } else if (s.prim === 'True') {
+      ret = true;
+    } else if (s.prim === 'False') {
+      ret = false;
+    }
+  } else if (Array.isArray(s)) {
+    const sc = s.length;
+    for (let i = 0; i < sc; i++) {
+      const n = me2(s[i]);
+      if (typeof n.key !== 'undefined') {
+        if (Array.isArray(ret)) {
+          ret = {
+            keys: [],
+            vals: [],
+          };
+        }
+        ret.keys.push(n.key);
+        ret.vals.push(n.val);
+      } else {
+        ret.push(n);
+      }
+    }
+  } else if (Object.prototype.hasOwnProperty.call(s, 'string')) {
+    ret = s.string;
+  } else if (Object.prototype.hasOwnProperty.call(s, 'int')) {
+    ret = parseInt(s.int, 10);
+  } else {
+    ret = s;
+  }
+  return ret;
+};
+
+utility.ml2mic = function me(mi: string): * {
+  const ret = [];
+  let inseq = false;
+  let seq = '';
+  let val = '';
+  let pl = 0;
+  let bl = 0;
+  let sopen = false;
+  let escaped = false;
+  for (let i = 0; i < mi.length; i++) {
+    if (val === '}' || val === ';') {
+      val = '';
+    }
+    if (inseq) {
+      if (mi[i] === '}') {
+        bl--;
+      } else if (mi[i] === '{') {
+        bl++;
+      }
+      if (bl === 0) {
+        const st = me(val);
+        ret.push({
+          prim: seq.trim(),
+          args: [st],
+        });
+        val = '';
+        bl = 0;
+        inseq = false;
+      }
+    } else if (mi[i] === '{') {
+      bl++;
+      seq = val;
+      val = '';
+      inseq = true;
+      continue;
+    } else if (escaped) {
+      val += mi[i];
+      escaped = false;
+      continue;
+    } else if ((i === (mi.length - 1) && sopen === false) || (mi[i] === ';' && pl === 0 && sopen === false)) {
+      if (i === (mi.length - 1)) val += mi[i];
+      if (val.trim() === '' || val.trim() === '}' || val.trim() === ';') {
+        val = '';
+        continue;
+      }
+      ret.push(utility.sexp2mic(val));
+      val = '';
+      continue;
+    } else if (mi[i] === '"' && sopen) {
+      sopen = false;
+    } else if (mi[i] === '"' && !sopen) {
+      sopen = true;
+    } else if (mi[i] === '\\') {
+      escaped = true;
+    } else if (mi[i] === '(') {
+      pl++;
+    } else if (mi[i] === ')') {
+      pl--;
+    }
+    val += mi[i];
+  }
+  return ret;
+};
+
+// Legacy commands
+utility.ml2tzjson = utility.sexp2mic;
+utility.tzjson2arr = utility.mic2arr;
+utility.mlraw2json = utility.ml2mic;
+utility.mintotz = utility.totez;
+utility.tztomin = utility.mutez;
+
+export default utility;
